tests: add performance test for sequential writes
*** Do not commit this ***

I'm sending the test because some people are interested in it.

This is the test I am using to diagnose sequential write performance
issue. The issue is visible when using a spinning disk. It writes
directly to vblk1, a second disk (see run.py change).

I get this output when running it:

0.469 Mb/s
0.312 Mb/s
0.323 Mb/s
0.354 Mb/s
0.163 Mb/s
0.100 Mb/s
0.388 Mb/s
0.293 Mb/s
0.401 Mb/s
Written 3.117 MB in 10.02 s

When doing sequential file writes on the host I get up to 70 Mb/s.

iostat shows saturated I/O, but svctm is high, 3-5 ms per request on
average, which looks like seek time is added. blktrace shows that
requests are indeed executed out of order. I'm investigating it
further.