random: collect entropy from interrupts
This is based on the FreeBSD implementation with slight
modifications. FreeBSD collects 2 bits of entropy from each interrupt,
based on:

 - interrupt vector number
 - current thread id
 - cycles counter

We collect 1 bit from the following information:

 - interrupt vector number
 - instruction pointer
 - cycles counter

For the reference I will describe what Linux is doing. Linux is using
the following information for entropy:

 - interrupt vector number
 - instruction pointer
 - cycles counter
 - jiffies counter

This information is hashed on Linux using a cryptographically unsecure
(but fast) hash into a per-CPU pool on every interrupt. Once every 64
interrupts or if a second has passed they feed the accumulated hash to
the common entropy pool and credit it for 1 bit of entropy.

I have verified that our solution is even remotely "sane" by
performing the following experiment. I ensured that interrupts are the
only source of entropy. I built a "tests" image. I ran OSv 2000 times
exeuting /tests/misc-random.so, which reads and prints 640 bytes from
/dev/random. It is the only thing which reads from /dev/random in the
system. I saved the outputs for comparison. Each output was unique
(there were no duplicates). It doesn't prove, however, that the
approach is secure.