Transparent dumpfile/tracefile usage + extended symbol resolution (if avail)
Instead of converting from trace dump files to tracefiles (basically raw trace data),
detect which we are reading on opening the file, and use the appropriate reader to extract data.
Also allow more chaining of symbol resolution, making it possible to use trace dump
symbol info when available (and debug symbol resolution fails).

(Note the dump format reader assumes the slightly revised dump symbol format, which in turn is in a
different patch. But since nothing until that patch actually creates SYMB chunks in dump files,
this is safe).

Signed-off-by: Calle Wilund <calle@cloudius-systems.com>