Optional module to add JVM JIT symbol resolution to trace dumps
If loaded into the JVM, will register with symbol callback and generate
symbol entries for all JIT:ed code currently running.

Does not deal with interpreter or dead/unloaded code. Note that HS call
frames are RBP-walkable (i.e. works with OSv backtrace), so as long as
we are in JIT code, we will get at least some symbols almost for free.

Signed-off-by: Calle Wilund <calle@cloudius-systems.com>